+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## [1.1.0] - 2021-03-24
+
+- Override `Struct#pretty_print` to keep redacted fields redacted when
+  pretty-printing via `pp`
+
 ## [1.0.0] - 2021-03-24
 
 - Initial release